Move URLSession helper to a separate file
