+++ Sources/ChangeLogGenerator/Fetcher/Fetcher.swift
@@ -103,32 +103,3 @@ struct Fetcher<Data: Decodable> {
         return try decoder.decode(Data.self, from: data)
     }
 }
-
-enum URLSessionError: Error {
-    case emptyData
-}
-
-extension URLSession {
-    public func dataAsync(for request: URLRequest) async throws -> (Data, URLResponse) {
-        if #available(iOS 15, macOS 12, *) {
-            return try await data(for: request)
-        }
-        else {
-            return try await withUnsafeThrowingContinuation { continuation in
-                let task = dataTask(with: request) { data, response, error in
-                    if let data = data, let response = response {
-                        continuation.resume(returning: (data, response))
-                    }
-                    else if let error = error {
-                        continuation.resume(throwing: error)
-                    }
-                    else {
-                        continuation.resume(throwing: URLSessionError.emptyData)
-                    }
-                }
-
-                task.resume()
-            }
-        }
-    }
-}
